Apply some fixes and improve package structure
