+++ utbot-spring-analyzer/src/main/java/application/TestApplicationConfiguration.java
@@ -1,8 +1,9 @@
-package analyzer;
+package application;
 
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import post_processors.UtBotBeanFactoryPostProcessor;
 
 @Configuration
 public class TestApplicationConfiguration {
