+++ utbot-spring-analyzer/src/main/java/utils/ResourceNames.java
@@ -1,0 +1,7 @@
+package utils;
+
+public class ResourceNames {
+
+    public static final String fakePropertiesFileName = "fake_application.properties";
+    public static final String fakeApplicationXmlFileName = "fake_application.xml";
+}
